% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_and_convert.R
\name{model_and_convert}
\alias{model_and_convert}
\title{The length of a string (in characters).}
\usage{
model_and_convert(fmla_list, svydesignObj, test.set, x.lab, y.lab, main)
}
\arguments{
\item{fmla_list}{a list of formulas from which to build svyglm objects}

\item{svydesignObj}{a svydesign object from which to creat a svyglm object}

\item{test.set}{a data frame whose variables match the svydesignObj, which is used to build ROC curves based on prediction}

\item{x.lab}{x-axis label for the ggplot}

\item{y.lab}{y-axis label for the ggplot}

\item{main}{title for the ggplot}
}
\value{
A data frame of ROC curve values for each input formula, and a ggplot of these values
}
\description{
The length of a string (in characters).
}
\seealso{
\code{\link{predROC}} which this function wraps
}

