% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaicSplom2.R
\name{mosaicSplom2}
\alias{mosaicSplom2}
\title{Create a grid of 2x2 survey-weighted mosaic plots}
\usage{
mosaicSplom2(factors = c(), mar_par = rep(0.5, 4))
}
\arguments{
\item{factors}{a vector of factor-type elements}

\item{mar_par}{a vector of length 2 controlling the margins of the plots}
}
\value{
A grid of 2x2 mosaics, one for each pair of the input factors
}
\description{
Create a grid of 2x2 survey-weighted mosaic plots
}
\examples{
factors <- c("DROWSY", "HEAVY_TRUCK",  "INT_HWY",  "SEX_IM", "SPEEDREL")
mosaicSplom2(factors)
}
