% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predROC.R
\name{predROC}
\alias{predROC}
\title{Helper function for WeightedROC::WeightedROC, to produce a data frame of ROC curve values based on a svyglm object and a test set}
\usage{
predROC(glm.obj, newData, plot = TRUE)
}
\arguments{
\item{glm.obj}{a svyglm object}

\item{newData}{a data frame whose variables are compatible with the svyglm object, for use in prediction}

\item{pot}{a boolean indicating whether or not to generate a plot}
}
\value{
A data frame of ROC curve values based on a svyglm object and a test set
}
\description{
Helper function for WeightedROC::WeightedROC, to produce a data frame of ROC curve values based on a svyglm object and a test set
}

