% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_and_convert.R
\name{model_and_convert}
\alias{model_and_convert}
\title{The length of a string (in characters).}
\usage{
model_and_convert(response, predictor_list, svydesignObj, test.set,
  plot = T)
}
\arguments{
\item{response}{a character string identifying the response variable}

\item{predictor_list}{a list whose elements are vectors of character strings specifying the desired predictor variables. A svyglm model is created for each element of the list.}

\item{svydesignObj}{a svydesign object from which to create a svyglm object}

\item{test.set}{a data frame whose variables match the svydesignObj, which is used to build ROC curves based on prediction}

\item{plot}{a Boolean indicating whether to plot the ROC curves for each model}
}
\value{
A data frame of ROC curve values for each input formula
}
\description{
The length of a string (in characters).
}
\examples{
data(list = c("GES2013.drivers.design", "GES2013.drivers"))
response <- "DROWSY"
predictor_list <- list(c ("HEAVY_TRUCK",  "INT_HWY",  "SEX_IM", "SPEEDREL"),
                                 c ("HEAVY_TRUCK",  "INT_HWY",  "SEX_IM"))
model_and_convert(response, predictor_list, GES2013.drivers.design, GES2013.drivers)
}
\seealso{
\code{\link{predROC}}
}
